Moved trovebox.api and trovebox.objects into packages
