+++ trovebox/__init__.py
@@ -4,9 +4,9 @@ __init__.py : Trovebox package top level
 from .http import Http
 from .errors import *
 from ._version import __version__
-from . import api_photo
-from . import api_tag
-from . import api_album
+from api import api_photo
+from api import api_tag
+from api import api_album
 
 LATEST_API_VERSION = 2
