+++ trovebox/api/__init__.py
@@ -1,0 +1,4 @@
+"""
+trovebox.api Package
+    Definitions for each of the Trovebox API endpoints
+"""
