+++ trovebox/objects/__init__.py
@@ -1,0 +1,4 @@
+"""
+trovebox.objects Package
+    Object classes returned by the API.
+"""
